followbutton from hook to component + address minor feedback
