+++ server/api/playbook_runs.go
@@ -700,19 +700,8 @@ func (h *PlaybookRunHandler) changeOwner(c *Context, w http.ResponseWriter, r *h
 		return
 	}
 
-	playbookRun, err := h.playbookRunService.GetPlaybookRun(vars["id"])
-	if err != nil {
-		h.HandleError(w, c.logger, err)
-		return
-	}
-
-	if !h.PermissionsCheck(w, c.logger, h.permissions.RunManageProperties(params.OwnerID, playbookRun.ID)) {
-		h.HandleErrorWithCode(w, c.logger, http.StatusForbidden, "target user is not a member of the run", err)
-		return
-	}
-
-	if !h.PermissionsCheck(w, c.logger, h.permissions.RunManageProperties(userID, playbookRun.ID)) {
-		h.HandleErrorWithCode(w, c.logger, http.StatusForbidden, "user has no permissions to change owner", err)
+	if !h.PermissionsCheck(w, c.logger, h.permissions.RunManageProperties(params.OwnerID, vars["id"])) {
+		h.HandleErrorWithCode(w, c.logger, http.StatusForbidden, "target user is not a member of the run", nil)
 		return
 	}
